pyproject.toml
[tool.poetry]
name = "aaaaAAAA"
version = "0.1.0"
description = "Our staff team submission for PyWeek 31. There will be ducks."
authors = ["Python Discord <info@pythondiscord.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
arcade = "^2.5.5"
pillow = "8.1.2"

[tool.poetry.dev-dependencies]
flake8 = "~=3.8"
flake8-annotations = "~=2.3"
flake8-bugbear = "~=20.1"
flake8-docstrings = "~=1.5"
flake8-import-order = "~=0.18"
flake8-string-format = "~=0.3"
flake8-tidy-imports = "~=4.1"
flake8-todo = "~=0.7"
pep8-naming = "~=0.11"
pre-commit = "~=2.1"
taskipy = "^1.6.0"

[tool.taskipy.tasks]
start = "python -m aaaaAAAA"
lint = "pre-commit run --all-files"
precommit = "pre-commit install"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
